Rename SetMultimapPropertyFactory.CodeGenerator -> SetMultimapProperty
SetMultimapPropertyFactory is now SetMultimapProperty.Factory
