+++ src/main/java/org/inferred/freebuilder/processor/Analyser.java
@@ -107,7 +107,7 @@ class Analyser {
       new MapProperty.Factory(),
       new MultisetProperty.Factory(),
       new ListMultimapProperty.Factory(),
-      new SetMultimapPropertyFactory(),
+      new SetMultimapProperty.Factory(),
       new OptionalProperty.Factory(),
       new BuildableProperty.Factory(),
       new DefaultProperty.Factory()); // Must be last, as it will always return a CodeGenerator
